Rename project to piperci-gman
