+++ .gitignore
@@ -131,4 +131,3 @@ dmypy.json
 *.db
 .idea/
 .envrc
-
