+++ tox.ini
@@ -26,7 +26,7 @@ max-complexity = 11
 
 [testenv:lint]
 deps = flake8>=3.6.0,<4
-commands = flake8 piedpiper_gman tests
+commands = flake8 piperci_gman tests
 skip_install = true
 usedevelop = false
 
@@ -36,4 +36,4 @@ usedevelop = true
 extras = test
 commands =
     pip freeze
-    pytest tests --cov=piedpiper_gman --cov-report=term-missing --no-cov-on-fail {posargs}
+    pytest tests --cov=piperci_gman --cov-report=term-missing --no-cov-on-fail {posargs}
